requirements.txt
websockets
soundcard
numpy<2.0.0
python-dotenv
aiohttp
requests
